delete thumbnail

+++ Util/deleteFile.js
@@ -1,19 +1,25 @@
 const fs = require('fs');
 
 const deleteSingleFile = (filePath) => {
-    fs.unlink(filePath, (err) => {
-        if (err) {
-            throw (err);
-        }
-    })
+    if (filePath) {
+        fs.unlink(filePath, (err) => {
+            if (err) {
+                throw (err);
+            }
+        })
+    }
 }
 
 const deleteMultiFile = (filePath) => {
-    filePath.map(path => fs.unlink(path, (err) => {
-        if (err) {
-            throw (err);
+    filePath.map(path => {
+        if (path) {
+            fs.unlink(path, (err) => {
+                if (err) {
+                    throw (err);
+                }
+            })
         }
-    }))
+    })
 }
 
 module.exports = {
